refactor aws package
